feat: LoginHome, LoginFooter로 컴포넌트 분리
